EDSC-4543: Move colormap to shared types; use zustand to pull out tags; update some tests with better patterns
