+++ static/src/js/components/Map/Map.tsx
@@ -948,9 +948,8 @@ const Map: React.FC<MapProps> = ({
       (control) => control instanceof LegendControl
     )
 
+    // Remove the legend control if not on the collection focused page and it exists
     if (!isFocusedCollectionPage && legendControl) {
-      // Remove the legend control if not on the collection focused page and it exists
-      console.log('removing legend control')
       controls.remove(legendControl)
     }
 
@@ -959,7 +958,6 @@ const Map: React.FC<MapProps> = ({
     // This helps ensure we have a reference to the legend control when we need to update it
     // otherwise issues would occur where the scrollbar would go to the top of the layer picker
     if (legendControl) {
-      console.log('update should be called')
       legendControl.update({
         collectionId: focusedCollectionId,
         imageryLayers
